MovieBench/requirement.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
asteroid-filterbanks==0.4.0
async-timeout==4.0.3
attrs==24.2.0
audioread==3.0.1
av==11.0.0
beautifulsoup4==4.12.3
bs4==0.0.2
cachetools==5.5.0
cffi==1.17.0
cinemagoer==2023.5.1
click==8.1.7
cmake==3.30.2
coloredlogs==15.0.1
colorlog==6.8.2
contourpy==1.1.1
ctranslate2==4.3.1
cycler==0.12.1
decorator==4.4.2
distro==1.9.0
docopt==0.6.2
docstring_parser==0.16
einops==0.8.0
exceptiongroup==1.2.2
faster-whisper==1.0.0
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
gcloud==0.18.3
google-ai-generativelanguage==0.1.0
google-api-core==2.19.1
google-auth==2.34.0
google-cloud-aiplatform==1.63.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.5
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-generativeai==0.1.0rc1
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpc-google-iam-v1==0.13.1
grpcio==1.66.0
grpcio-status==1.62.3
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.24.5
humanfriendly==10.0
HyperPyYAML==1.2.2
imageio==2.35.0
imageio-ffmpeg==0.5.1
IMDbPY==2022.7.9
importlib_resources==6.4.0
jiter==0.5.0
joblib==1.4.2
julius==0.2.7
kiwisolver==1.4.5
lazy_loader==0.4
librosa==0.10.2.post1
lightning==2.3.3
lightning-utilities==0.11.6
lit==18.1.8
llvmlite==0.41.1
lxml==5.3.0
Mako==1.3.5
markdown-it-py==3.0.0
matplotlib==3.7.5
mdurl==0.1.2
mkl-service==2.4.0
moviepy==1.0.3
msgpack==1.0.8
multidict==6.0.5
nltk==3.8.2
numba==0.58.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
oauth2client==4.1.3
omegaconf==2.3.0
onnxruntime==1.18.1
openai==1.40.6
opencv-python==4.10.0.84
optuna==3.6.1
pandas==2.0.3
pooch==1.8.2
primePy==1.3
proglog==0.1.10
proto-plus==1.24.0
protobuf==4.25.4
pyannote.audio==3.1.1
pyannote.core==5.0.0
pyannote.database==5.1.0
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
pyparsing==3.1.2
pytorch-lightning==2.3.3
pytorch-metric-learning==2.6.0
pytz==2024.1
regex==2024.7.24
rich==13.7.1
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safetensors==0.4.4
scikit-learn==1.3.2
scipy==1.10.1
semver==3.0.2
sentencepiece==0.2.0
shapely==2.0.6
shellingham==1.5.4
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
speechbrain==1.0.0
SQLAlchemy==2.0.32
tabulate==0.9.0
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.1.2
torch-audiomentations==0.11.1
torch-pitch-shift==1.2.4
torchaudio==2.0.1
torchmetrics==1.4.1
torchvision==0.15.1
tqdm==4.66.5
transformers==4.40.0
triton==2.0.0
typer==0.12.3
tzdata==2024.1
yarl==1.9.4
